Moved plugin-sidebar error-boundary to separate file
